trying to use the hostname command from a sandboxed image.
when used from a sif file, the hostname reported back is a commit hash, but this isnt useful in tracing commands back to a particular node. However, I am unable to bind into a sandboxed image to read /etc/hostname. Trying the hostname command from sandbox to see if this might resolve my issue.